hotfix: behave correctly when points with times larger than the scan period exist
GLIM cuts IMU sequences according to the period of LiDAR scans.
For some LiDAR configurations, it might happen that the span of point
timestamps in one scan is larger than its publishing period.

For example, a 360deg Hesai LiDAR decoded with Nebula will have points
from a span of up to 120ms for a 10Hz frequency (100ms period), due to
angle and fire time correction.

Signed-off-by: Max SCHMELLER <max.schmeller@tier4.jp>